# os-manager-api/app/Http/Controllers/OrdemServicoController.php
<?php

namespace App\Http\Controllers;

use App\Models\OrdemServico;
use App\Models\Status;
use App\Http\Requests\OrdemServicoRequest;
use Illuminate\Http\Request;
use Illuminate\Support\Str;
use OpenApi\Attributes as OA;

#[OA\Tag(name: "OrdensServico", description: "Endpoints para gerenciamento de ordens de serviço")]
class OrdemServicoController extends Controller
{
   #[OA\Get(
        path: "/api/ordens",
        tags: ["Ordens de Servico"],
        summary: "Lista todas as ordens de serviço com filtros",
        security: [["bearerAuth" => []]],
        parameters: [
            new OA\Parameter(
                name: "status",
                in: "query",
                description: "Filtrar por nome do status",
                required: false,
                schema: new OA\Schema(type: "string")
            ),
            new OA\Parameter(
                name: "categoria",
                in: "query",
                description: "Filtrar por nome da categoria",
                required: false,
                schema: new OA\Schema(type: "string")
            ),
            new OA\Parameter(
                name: "busca",
                in: "query",
                description: "Busca por título ou descrição",
                required: false,
                schema: new OA\Schema(type: "string")
            ),
            new OA\Parameter(
                name: "ID",
                in: "query",
                description: "Filtrar por ID (UUID) ou código de rastreio",
                required: false,
                schema: new OA\Schema(type: "string") 
            )
        ],
        responses: [
            new OA\Response(response: 200, description: "Lista de ordens recuperada com sucesso"),
            new OA\Response(response: 401, description: "Não autenticado"),
            new OA\Response(response: 404, description: "Ordens de serviço não encontradas")
        ]
    )]
   public function index(Request $request)
    {
        $query = OrdemServico::with([
            'usuario', 'tecnico',
            'status', 'categoria',
            'urgencia', 'prioridade'
        ]);

        $statusAtivo = $request->has('ativo')
            ? filter_var($request->ativo, FILTER_VALIDATE_BOOLEAN)
            : true;

        $query->where('ativo', $statusAtivo);

        $buscaId = $request->input('id') ?? $request->input('ID');

        if (!empty($buscaId)) {
            $query->where(function ($q) use ($buscaId) {
                // CÓDIGO ANTIGO:
                /*
                if (Str::isUuid($buscaId)) {
                    // Se for UUID, busca nas colunas de UUID
                    $q->where('id', $buscaId)
                      ->orWhere('codigo_rastreio', $buscaId);
                } else {
                    // Se não for UUID (ex: "24"), evita erro 22P02 do Postgres 
                    // Se codigo_rastreio for string comum, busca aqui. 
                    // Se for UUID, o whereRaw('1=0') impede o crash.
                    $q->whereRaw('1=0'); 
                }
                */

                // CÓDIGO NOVO CORRIGIDO:
                if (Str::isUuid($buscaId)) {
                    $q->where('codigo_rastreio', $buscaId);
                } elseif (is_numeric($buscaId)) {
                    $q->where('id', $buscaId);
                } else {
                    $q->where('codigo_rastreio', $buscaId);
                }
            });
        }

        // Filtros por relacionamentos (Mantendo sua estrutura original)
        if ($request->filled('status')) {
            $query->whereHas('status', fn($q) => $q->where('nome', $request->status));
        }

        if ($request->filled('categoria')) {
            $query->whereHas('categoria', fn($q) => $q->where('nome', $request->categoria));
        }

        if ($request->filled('urgencia')) {
            $query->whereHas('urgencia', fn($q) => $q->where('nome', $request->urgencia));
        }

        if ($request->filled('prioridade')) {
            $query->whereHas('prioridade', fn($q) => $q->where('nome', $request->prioridade));
        }

        if ($request->filled('tecnico_id')) {
            $query->where('tecnico_id', $request->tecnico_id);
        }

        if ($request->filled('busca')) {
            $query->where(function ($q) use ($request) {
                $q->where('titulo', 'ilike', '%' . $request->busca . '%')
                  ->orWhere('descricao', 'ilike', '%' . $request->busca . '%');
            });
        }

        return response()->json(
            $query->orderBy('criado_em', 'desc')->get(),
            200
        );
    }

    #[OA\Post(
        path: "/api/ordens",
        tags: ["Ordens de Servico"],
        summary: "Cria uma nova ordem de serviço",
        security: [["bearerAuth" => []]],
        requestBody: new OA\RequestBody(
            required: true,
            content: new OA\JsonContent(
                properties: [
                    new OA\Property(property: "titulo", type: "string", example: "Estagiário esqueceu a senha"),
                    new OA\Property(property: "descricao", type: "string", example: "O estagiário João da Silva esqueceu a senha do sistema e precisa de ajuda para resetar o acesso."),
                    new OA\Property(property: "localizacao", type: "string", example: "Bloco A, Sala 102"),
                    new OA\Property(property: "categoria", type: "string", example: "Acesso"),
                    new OA\Property(property: "urgencia", type: "string", example: "Alta"),
                    new OA\Property(property: "prioridade", type: "string", example: "Muito Alta")
                ]
            )
        ),
        responses: [
            new OA\Response(response: 201, description: "Ordem de serviço criada com sucesso"),
            new OA\Response(response: 422, description: "Erro de validação nos dados enviados"),
            new OA\Response(response: 401, description: "Não autenticado")
        ]
    )]
    public function store(OrdemServicoRequest $request)
    {
        $usuarioLogado = $request->user();
        $idDonoDoChamado = $usuarioLogado->id;
        $cargoId = $usuarioLogado->cargo_id;
        $cargosPrivilegiados = [1, 2]; 

        if (in_array($cargoId, $cargosPrivilegiados) && $request->filled('usuario_id')) {
            $idDonoDoChamado = $request->usuario_id;
        }

        $statusNovoId = Status::where('nome', 'Novo')->value('id');

        $novaOrdem = OrdemServico::create([
            'titulo'        => $request->titulo,
            'descricao'     => $request->descricao,
            'usuario_id'    => $idDonoDoChamado,
            'categoria_id'  => $request->categoria_id,
            'localizacao'   => $request->localizacao,
            'status_id'     => $request->status_id ?? $statusNovoId,
            'urgencia_id'   => $request->urgencia_id,
            'prioridade_id' => $request->prioridade_id,
            'ativo'         => true,
        ]);

        return response()->json(
            $novaOrdem->load(['usuario', 'tecnico', 'status', 'categoria', 'urgencia', 'prioridade']),
            201
        );
    }

   #[OA\Get(
        path: "/api/ordens/{id}",
        tags: ["Ordens de Servico"],
        summary: "Exibe detalhes de uma ordem de serviço específica",
        description: "Busca uma Ordem de Serviço pelo ID (UUID) ou pelo Código de Rastreio",
        security: [["bearerAuth" => []]],
        parameters: [
            new OA\Parameter(
                name: "id",
                in: "path",
                required: true,
                description: "ID (UUID) ou Código de Rastreio",
                schema: new OA\Schema(type: "string")
            )
        ],
        responses: [
            new OA\Response(response: 200, description: "Dados da ordem de serviço"),
            new OA\Response(response: 401, description: "Não autenticado"),
            new OA\Response(response: 404, description: "Ordem de serviço não encontrada")
        ]
    )]
    public function show($id)
    {
        $query = OrdemServico::with([
            'usuario', 'tecnico',
            'status', 'categoria',
            'urgencia', 'prioridade'
        ]);

        $ordem = $query->where(function ($q) use ($id) {
            // CÓDIGO ANTIGO:
            /*
            if (Str::isUuid($id)) {
                $q->where('id', $id)->orWhere('codigo_rastreio', $id);
            } else {
                $q->whereRaw('1=0');
            }
            */

            // CÓDIGO NOVO CORRIGIDO:
            if (Str::isUuid($id)) {
                $q->where('codigo_rastreio', $id);
            } elseif (is_numeric($id)) {
                $q->where('id', $id);
            } else {
                $q->where('codigo_rastreio', $id);
            }
        })->first();

        if (!$ordem) {
            return response()->json(['message' => 'Ordem de serviço não encontrada'], 404);
        }

        return response()->json($ordem, 200);
    }

     #[OA\Put(
        path: "/api/ordens/{id}",
        tags: ["Ordens de Servico"],
        summary: "Atualiza uma ordem de serviço",
        security: [["bearerAuth" => []]],
        parameters: [
            new OA\Parameter(name: "id", in: "path", required: true, schema: new OA\Schema(type: "string"))
        ],
        requestBody: new OA\RequestBody(
            required: true,
            description: "Dados para atualização da Ordem de Serviço",
            content: new OA\MediaType(
                mediaType: "application/json",
                schema: new OA\Schema(
                    type: "object",
                    properties: [
                        new OA\Property(property: "status_id", type: "integer", example: 2),
                        new OA\Property(property: "urgencia_id", type: "integer", example: 1),
                        new OA\Property(property: "prioridade_id", type: "integer", example: 3),
                        new OA\Property(property: "tecnico_id", type: "integer", example: 5),
                        new OA\Property(property: "motivo_pausa", type: "string", nullable: true, example: "Aguardando peça"),
                        new OA\Property(property: "solucao", type: "string", nullable: true, example: "Resolvido")
                    ]
                )
            )
        ),
        responses: [
            new OA\Response(response: 200, description: "Ordem de serviço atualizada"),
            new OA\Response(response: 403, description: "Acesso negado"),
            new OA\Response(response: 404, description: "Ordem de serviço não encontrada")
        ]
    )]
    public function update(OrdemServicoRequest $request, $id)
    {
        $item = OrdemServico::where(function ($q) use ($id) {
            // CÓDIGO ANTIGO:
            /*
            if (Str::isUuid($id)) {
                $q->where('id', $id)->orWhere('codigo_rastreio', $id);
            } else {
                $q->whereRaw('1=0');
            }
            */

            // CÓDIGO NOVO CORRIGIDO:
            if (Str::isUuid($id)) {
                $q->where('codigo_rastreio', $id);
            } elseif (is_numeric($id)) {
                $q->where('id', $id);
            } else {
                $q->where('codigo_rastreio', $id);
            }
        })->firstOrFail();

        $dados = $request->only([
            'status_id',
            'urgencia_id',
            'prioridade_id',
            'tecnico_id',
            'motivo_pausa',
            'solucao'
        ]);

        // Lógica de pausa original completa
        $statusAntigo = $item->status?->nome;
        $statusNovo = $request->status ?? $item->status?->nome;
        $estadosPausa = ['Pausado', 'Aguardando Peça'];

        if (in_array($statusNovo, $estadosPausa)) {
            if (!in_array($statusAntigo, $estadosPausa)) {
                $dados['pausado_em'] = now();
            }
        } else {
            if (in_array($statusAntigo, $estadosPausa) && $item->pausado_em) {
                $minutos = now()->diffInMinutes($item->pausado_em);
                $dados['tempo_pausado_minutos'] = ($item->tempo_pausado_minutos ?? 0) + $minutos;
                $dados['pausado_em'] = null;
                $dados['motivo_pausa'] = null;
            }
        }

        $item->update($dados);

        return response()->json(
            $item->load(['usuario', 'tecnico', 'status', 'categoria', 'urgencia', 'prioridade']), 
            200
        );
    }

    #[OA\Delete(
        path: "/api/ordens/{id}",
        tags: ["Ordens de Servico"],
        summary: "Remove (desativa) uma ordem de serviço",
        security: [["bearerAuth" => []]],
        parameters: [
            new OA\Parameter(name: "id", in: "path", required: true, schema: new OA\Schema(type: "string"))
        ],
        responses: [
            new OA\Response(response: 200, description: "Ordem de serviço removida"),
            new OA\Response(response: 403, description: "Acesso negado")
        ]
    )]
    public function destroy($id)
    {
        $item = OrdemServico::where(function ($q) use ($id) {
            // CÓDIGO ANTIGO:
            /*
            if (Str::isUuid($id)) {
                $q->where('id', $id)->orWhere('codigo_rastreio', $id);
            } else {
                $q->whereRaw('1=0');
            }
            */

            // CÓDIGO NOVO CORRIGIDO:
            if (Str::isUuid($id)) {
                $q->where('codigo_rastreio', $id);
            } elseif (is_numeric($id)) {
                $q->where('id', $id);
            } else {
                $q->where('codigo_rastreio', $id);
            }
        })->firstOrFail();
            
        $item->update(['ativo' => false]);

        return response()->json(['message' => 'Enviado para a lixeira com sucesso'], 200);
    }
}