# os-manager-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OS Manager API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/ordens": {
            "get": {
                "tags": [
                    "Ordens de Servico"
                ],
                "summary": "Lista todas as ordens de serviço com filtros",
                "operationId": "e9954d9a37630cdf929ae616b77d4103",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por nome do status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "categoria",
                        "in": "query",
                        "description": "Filtrar por nome da categoria",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "busca",
                        "in": "query",
                        "description": "Busca por título ou descrição",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ID",
                        "in": "query",
                        "description": "Filtrar por ID (UUID) ou código de rastreio",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de ordens recuperada com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Ordens de serviço não encontradas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ordens de Servico"
                ],
                "summary": "Cria uma nova ordem de serviço",
                "operationId": "57a9b14a3f752b7e724fb3443663c8e5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Estagiário esqueceu a senha"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "O estagiário João da Silva esqueceu a senha do sistema e precisa de ajuda para resetar o acesso."
                                    },
                                    "localizacao": {
                                        "type": "string",
                                        "example": "Bloco A, Sala 102"
                                    },
                                    "categoria": {
                                        "type": "string",
                                        "example": "Acesso"
                                    },
                                    "urgencia": {
                                        "type": "string",
                                        "example": "Alta"
                                    },
                                    "prioridade": {
                                        "type": "string",
                                        "example": "Muito Alta"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ordem de serviço criada com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação nos dados enviados"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/ordens/{id}": {
            "get": {
                "tags": [
                    "Ordens de Servico"
                ],
                "summary": "Exibe detalhes de uma ordem de serviço específica",
                "description": "Busca uma Ordem de Serviço pelo ID (UUID) ou pelo Código de Rastreio",
                "operationId": "47855ee63438258517219d38894ca714",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID (UUID) ou Código de Rastreio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados da ordem de serviço"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Ordem de serviço não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Ordens de Servico"
                ],
                "summary": "Atualiza uma ordem de serviço",
                "operationId": "a5d619b2ba267041732114d8f319a67f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados para atualização da Ordem de Serviço",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "urgencia_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "prioridade_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "tecnico_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "motivo_pausa": {
                                        "type": "string",
                                        "example": "Aguardando peça",
                                        "nullable": true
                                    },
                                    "solucao": {
                                        "type": "string",
                                        "example": "Resolvido",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ordem de serviço atualizada"
                    },
                    "403": {
                        "description": "Acesso negado"
                    },
                    "404": {
                        "description": "Ordem de serviço não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Ordens de Servico"
                ],
                "summary": "Remove (desativa) uma ordem de serviço",
                "operationId": "3ddb628fff4f4f9ebafbcbbc20f22bd1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ordem de serviço removida"
                    },
                    "403": {
                        "description": "Acesso negado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}